two classed moved to another directory
